feat: add userdata to metadata API
